[sca/otbn] Add code to capture and analyze modular inverse OTBN snippet
Extends the otbn_vertical capture functions with code to capture traces
of modular inversion computation on OTBN.
For this, the OTBN app to be captured has to be configured in the
capture config yaml file.
Further, a new TVLA mode 'otbn_modinv' is introduced, to enable/handle
t-test analysis for modular inversion traces.

Signed-off-by: Moritz Wettermann <moritz.wettermann@gi-de.com>